fix issue with settlement rollback when some trades succeed and need to be resubmitted
