+++ integrationtests/src/test/kotlin/co/chainring/integrationtests/settlement/SettlementFailureTest.kt
@@ -1,0 +1,986 @@
+package co.chainring.integrationtests.settlement
+
+import co.chainring.apps.api.model.websocket.LastTrade
+import co.chainring.apps.api.model.websocket.LastTradeDirection
+import co.chainring.apps.api.model.websocket.OrderBook
+import co.chainring.apps.api.model.websocket.OrderBookEntry
+import co.chainring.apps.api.model.websocket.SubscriptionTopic
+import co.chainring.apps.api.model.websocket.TradeUpdated
+import co.chainring.core.model.db.ExecutionRole
+import co.chainring.core.model.db.OrderSide
+import co.chainring.core.model.db.OrderStatus
+import co.chainring.core.model.db.SettlementStatus
+import co.chainring.core.model.db.WithdrawalEntity
+import co.chainring.core.model.db.WithdrawalStatus
+import co.chainring.integrationtests.testutils.AppUnderTestRunner
+import co.chainring.integrationtests.testutils.OrderBaseTest
+import co.chainring.integrationtests.testutils.waitForFinalizedWithdrawal
+import co.chainring.integrationtests.utils.AssetAmount
+import co.chainring.integrationtests.utils.ExpectedBalance
+import co.chainring.integrationtests.utils.assertAmount
+import co.chainring.integrationtests.utils.assertBalancesMessageReceived
+import co.chainring.integrationtests.utils.assertContainsBalancesMessage
+import co.chainring.integrationtests.utils.assertContainsLimitsMessage
+import co.chainring.integrationtests.utils.assertContainsTradeUpdatedMessage
+import co.chainring.integrationtests.utils.assertLimitOrderCreatedMessageReceived
+import co.chainring.integrationtests.utils.assertLimitsMessageReceived
+import co.chainring.integrationtests.utils.assertMarketOrderCreatedMessageReceived
+import co.chainring.integrationtests.utils.assertMessagesReceived
+import co.chainring.integrationtests.utils.assertOrderBookMessageReceived
+import co.chainring.integrationtests.utils.assertOrderUpdatedMessageReceived
+import co.chainring.integrationtests.utils.assertPricesMessageReceived
+import co.chainring.integrationtests.utils.assertTradeCreatedMessageReceived
+import org.jetbrains.exposed.sql.transactions.transaction
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.extension.ExtendWith
+import java.math.BigDecimal
+import kotlin.test.Test
+
+@ExtendWith(AppUnderTestRunner::class)
+class SettlementFailureTest : OrderBaseTest() {
+
+    @Test
+    fun `settlement failure - all trades in batch fail`() {
+        val market = btcEthMarket
+        val baseSymbol = btc
+        val quoteSymbol = eth
+
+        val (takerApiClient, takerWallet, takerWsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.5"),
+                AssetAmount(quoteSymbol, "2"),
+            ),
+            deposits = listOf(
+                AssetAmount(baseSymbol, "0.1"),
+                AssetAmount(quoteSymbol, "2"),
+            ),
+        )
+
+        val (makerApiClient, makerWallet, makerWsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.5"),
+                AssetAmount(quoteSymbol, "2"),
+            ),
+            deposits = listOf(
+                AssetAmount(quoteSymbol, "2"),
+            ),
+        )
+
+        // starting onchain balances
+        val makerStartingBaseBalance = makerWallet.getExchangeBalance(baseSymbol)
+        val makerStartingQuoteBalance = makerWallet.getExchangeBalance(quoteSymbol)
+        val takerStartingBaseBalance = takerWallet.getExchangeBalance(baseSymbol)
+        val takerStartingQuoteBalance = takerWallet.getExchangeBalance(quoteSymbol)
+
+        val baseWithdrawalAmount = AssetAmount(baseSymbol, "0.015")
+
+        val pendingBaseWithdrawal = takerApiClient.createWithdrawal(takerWallet.signWithdraw(baseSymbol.name, baseWithdrawalAmount.inFundamentalUnits)).withdrawal
+        assertEquals(WithdrawalStatus.Pending, pendingBaseWithdrawal.status)
+
+        takerWsClient.apply {
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = takerStartingBaseBalance, available = takerStartingBaseBalance - baseWithdrawalAmount),
+                    ExpectedBalance(quoteSymbol, total = takerStartingQuoteBalance, available = takerStartingQuoteBalance),
+                ),
+            )
+            assertLimitsMessageReceived(market, base = takerStartingBaseBalance - baseWithdrawalAmount, quote = takerStartingQuoteBalance)
+        }
+
+        waitForFinalizedWithdrawal(pendingBaseWithdrawal.id)
+
+        takerWsClient.apply {
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = takerStartingBaseBalance - baseWithdrawalAmount, available = takerStartingBaseBalance - baseWithdrawalAmount),
+                    ExpectedBalance(quoteSymbol, total = takerStartingQuoteBalance, available = takerStartingQuoteBalance),
+                ),
+            )
+        }
+
+        // hack to resubmit the same transaction again bypassing the sequencer. This way taker will not have a
+        // sufficient on chain balance for the order so settlement will fail.
+        transaction {
+            WithdrawalEntity[pendingBaseWithdrawal.id].status = WithdrawalStatus.Sequenced
+        }
+
+        waitForFinalizedWithdrawal(pendingBaseWithdrawal.id)
+
+        takerWsClient.apply {
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = takerStartingBaseBalance - baseWithdrawalAmount * BigDecimal(2), available = takerStartingBaseBalance - baseWithdrawalAmount),
+                    ExpectedBalance(quoteSymbol, total = takerStartingQuoteBalance, available = takerStartingQuoteBalance),
+                ),
+            )
+        }
+
+        // place a limit order
+        val limitBuyOrderApiResponse = makerApiClient.createLimitOrder(
+            market,
+            OrderSide.Buy,
+            amount = BigDecimal("0.08"),
+            price = BigDecimal("17.55"),
+            makerWallet,
+        )
+        makerWsClient.apply {
+            assertLimitOrderCreatedMessageReceived(limitBuyOrderApiResponse)
+            assertLimitsMessageReceived(market, base = BigDecimal("0"), quote = BigDecimal("0.596"))
+        }
+
+        listOf(makerWsClient, takerWsClient).forEach { wsClient ->
+            wsClient.assertOrderBookMessageReceived(
+                market.id,
+                OrderBook(
+                    marketId = market.id,
+                    buy = listOf(
+                        OrderBookEntry(price = "17.550", size = "0.08".toBigDecimal()),
+                    ),
+                    sell = emptyList(),
+                    last = LastTrade("0.000", LastTradeDirection.Unchanged),
+                ),
+            )
+        }
+
+        // place a sell order
+        val marketBuyOrderApiResponse = takerApiClient.createMarketOrder(
+            market,
+            OrderSide.Sell,
+            BigDecimal("0.08"),
+            takerWallet,
+        )
+
+        takerWsClient.apply {
+            assertMarketOrderCreatedMessageReceived(marketBuyOrderApiResponse)
+            assertTradeCreatedMessageReceived(
+                order = marketBuyOrderApiResponse,
+                price = BigDecimal("17.55"),
+                amount = AssetAmount(baseSymbol, marketBuyOrderApiResponse.order.amount),
+                fee = AssetAmount(quoteSymbol, "0.02808"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(1, msg.order.executions.size)
+                assertAmount(AssetAmount(baseSymbol, "0.08"), msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "17.550"), msg.order.executions[0].price)
+                assertEquals(ExecutionRole.Taker, msg.order.executions[0].role)
+            }
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = BigDecimal("0.07"), available = BigDecimal("0.005")),
+                    ExpectedBalance(quoteSymbol, total = BigDecimal("2.0"), available = BigDecimal("3.37592")),
+                ),
+            )
+        }
+
+        makerWsClient.apply {
+            assertTradeCreatedMessageReceived(
+                order = limitBuyOrderApiResponse,
+                price = BigDecimal("17.55"),
+                amount = AssetAmount(baseSymbol, "0.08"),
+                fee = AssetAmount(quoteSymbol, "0.01404"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(1, msg.order.executions.size)
+                assertAmount(AssetAmount(baseSymbol, "0.08"), msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "17.550"), msg.order.executions[0].price)
+                assertEquals(ExecutionRole.Maker, msg.order.executions[0].role)
+            }
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = BigDecimal("0"), available = BigDecimal("0.08")),
+                    ExpectedBalance(quoteSymbol, total = BigDecimal("2.0"), available = BigDecimal("0.58196")),
+                ),
+            )
+        }
+
+        listOf(makerWsClient, takerWsClient).forEach { wsClient ->
+            wsClient.assertOrderBookMessageReceived(
+                market.id,
+                OrderBook(
+                    marketId = market.id,
+                    buy = listOf(),
+                    sell = listOf(),
+                    last = LastTrade("17.550", LastTradeDirection.Up),
+                ),
+            )
+        }
+        val trade = getTradesForOrders(listOf(marketBuyOrderApiResponse.orderId)).first()
+
+        takerWsClient.apply {
+            assertPricesMessageReceived(market.id)
+            assertLimitsMessageReceived(market, base = BigDecimal("0.005"), quote = BigDecimal("3.37592"))
+        }
+
+        makerWsClient.apply {
+            assertPricesMessageReceived(market.id)
+            assertLimitsMessageReceived(market, base = BigDecimal("0.08"), quote = BigDecimal("0.58196"))
+        }
+
+        assertAmount(AssetAmount(baseSymbol, "0.08"), trade.amount)
+        assertAmount(AssetAmount(quoteSymbol, "17.550"), trade.price)
+
+        waitForSettlementToFinish(listOf(trade.id.value), SettlementStatus.Failed)
+
+        takerWsClient.apply {
+            assertMessagesReceived(3) { messages ->
+                assertContainsTradeUpdatedMessage(messages) { msg ->
+                    assertEquals(marketBuyOrderApiResponse.orderId, msg.trade.orderId)
+                    assertEquals(SettlementStatus.Failed, msg.trade.settlementStatus)
+                    assertEquals("Insufficient Balance", msg.trade.error)
+                }
+                assertContainsBalancesMessage(
+                    messages,
+                    listOf(
+                        ExpectedBalance(baseSymbol, total = AssetAmount(baseSymbol, "0.07"), available = takerStartingBaseBalance - baseWithdrawalAmount),
+                        ExpectedBalance(takerStartingQuoteBalance),
+                    ),
+                )
+                assertContainsLimitsMessage(messages, market, base = takerStartingBaseBalance - baseWithdrawalAmount, quote = takerStartingQuoteBalance)
+            }
+        }
+
+        makerWsClient.apply {
+            assertMessagesReceived(3) { messages ->
+                assertContainsTradeUpdatedMessage(messages) { msg ->
+                    assertEquals(limitBuyOrderApiResponse.orderId, msg.trade.orderId)
+                    assertEquals(SettlementStatus.Failed, msg.trade.settlementStatus)
+                    assertEquals("Insufficient Balance", msg.trade.error)
+                }
+                assertContainsBalancesMessage(
+                    messages,
+                    listOf(
+                        ExpectedBalance(makerStartingBaseBalance),
+                        ExpectedBalance(makerStartingQuoteBalance),
+                    ),
+                )
+                assertContainsLimitsMessage(messages, market, base = makerStartingBaseBalance, quote = makerStartingQuoteBalance)
+            }
+        }
+
+        makerWsClient.close()
+    }
+
+    @Test
+    fun `settlement failure - some trades in batch fail, some succeed`() {
+        val market = btcEthMarket
+        val baseSymbol = btc
+        val quoteSymbol = eth
+
+        val (takerApiClient, takerWallet, takerWsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.5"),
+                AssetAmount(quoteSymbol, "10"),
+            ),
+            deposits = listOf(
+                AssetAmount(quoteSymbol, "10"),
+            ),
+            subscribeToOrderBook = false,
+            subscribeToOrderPrices = false,
+        )
+
+        val (makerApiClient, makerWallet, makerWsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.1"),
+            ),
+            deposits = listOf(
+                AssetAmount(baseSymbol, "0.05"),
+            ),
+            subscribeToOrderBook = false,
+            subscribeToOrderPrices = false,
+        )
+
+        val (maker2ApiClient, maker2Wallet, maker2WsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.1"),
+            ),
+            deposits = listOf(
+                AssetAmount(baseSymbol, "0.03"),
+            ),
+            subscribeToOrderBook = false,
+            subscribeToOrderPrices = false,
+        )
+
+        // starting onchain balances
+        val makerStartingBaseBalance = makerWallet.getExchangeBalance(baseSymbol)
+        val makerStartingQuoteBalance = makerWallet.getExchangeBalance(quoteSymbol)
+        val maker2StartingBaseBalance = maker2Wallet.getExchangeBalance(baseSymbol)
+        val baseSymbolMakerTradeAmount = AssetAmount(baseSymbol, "0.03")
+        val makerNotionalMinusFee = AssetAmount(quoteSymbol, "0.51975")
+
+        val baseWithdrawalAmount = AssetAmount(baseSymbol, "0.02")
+
+        val pendingBaseWithdrawal = makerApiClient.createWithdrawal(makerWallet.signWithdraw(baseSymbol.name, baseWithdrawalAmount.inFundamentalUnits)).withdrawal
+        assertEquals(WithdrawalStatus.Pending, pendingBaseWithdrawal.status)
+
+        makerWsClient.apply {
+            assertBalancesMessageReceived()
+            assertLimitsMessageReceived(market, base = makerStartingBaseBalance - baseWithdrawalAmount, quote = makerStartingQuoteBalance)
+        }
+
+        waitForFinalizedWithdrawal(pendingBaseWithdrawal.id)
+
+        makerWsClient.apply {
+            assertBalancesMessageReceived()
+        }
+
+        // hack to resubmit the same transaction again bypassing the sequencer. This way taker will not have a
+        // sufficient on chain balance for the order so settlement will fail.
+        transaction {
+            WithdrawalEntity[pendingBaseWithdrawal.id].status = WithdrawalStatus.Sequenced
+        }
+
+        waitForFinalizedWithdrawal(pendingBaseWithdrawal.id)
+
+        makerWsClient.apply {
+            assertBalancesMessageReceived()
+        }
+
+        // place a limit order
+        val limitSellOrderApiResponse = makerApiClient.createLimitOrder(
+            market,
+            OrderSide.Sell,
+            amount = BigDecimal("0.03"),
+            price = BigDecimal("17.50"),
+            makerWallet,
+        )
+        makerWsClient.apply {
+            assertLimitOrderCreatedMessageReceived(limitSellOrderApiResponse)
+            assertLimitsMessageReceived(market.id)
+        }
+
+        val limitSellOrder2ApiResponse = maker2ApiClient.createLimitOrder(
+            market,
+            OrderSide.Sell,
+            amount = BigDecimal("0.03"),
+            price = BigDecimal("17.50"),
+            maker2Wallet,
+        )
+        maker2WsClient.apply {
+            assertLimitOrderCreatedMessageReceived(limitSellOrder2ApiResponse)
+            assertLimitsMessageReceived(market.id)
+        }
+
+        // place a sell order
+        val marketSellOrderApiResponse = takerApiClient.createMarketOrder(
+            market,
+            OrderSide.Buy,
+            BigDecimal("0.06"),
+            takerWallet,
+        )
+
+        takerWsClient.apply {
+            assertMarketOrderCreatedMessageReceived(marketSellOrderApiResponse)
+            assertTradeCreatedMessageReceived(
+                order = marketSellOrderApiResponse,
+                price = BigDecimal("17.50"),
+                amount = baseSymbolMakerTradeAmount,
+                fee = AssetAmount(quoteSymbol, "0.0105"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertTradeCreatedMessageReceived(
+                order = marketSellOrderApiResponse,
+                price = BigDecimal("17.50"),
+                amount = baseSymbolMakerTradeAmount,
+                fee = AssetAmount(quoteSymbol, "0.0105"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(2, msg.order.executions.size)
+                assertAmount(baseSymbolMakerTradeAmount, msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "17.500"), msg.order.executions[0].price)
+                assertAmount(baseSymbolMakerTradeAmount, msg.order.executions[1].amount)
+                assertAmount(AssetAmount(quoteSymbol, "17.500"), msg.order.executions[1].price)
+                assertEquals(ExecutionRole.Taker, msg.order.executions[0].role)
+                assertEquals(ExecutionRole.Taker, msg.order.executions[1].role)
+            }
+            assertBalancesMessageReceived()
+        }
+
+        makerWsClient.apply {
+            assertTradeCreatedMessageReceived(
+                order = limitSellOrderApiResponse,
+                price = BigDecimal("17.50"),
+                amount = baseSymbolMakerTradeAmount,
+                fee = AssetAmount(quoteSymbol, "0.00525"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(1, msg.order.executions.size)
+                assertAmount(AssetAmount(baseSymbol, "0.03"), msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "17.500"), msg.order.executions[0].price)
+                assertEquals(ExecutionRole.Maker, msg.order.executions[0].role)
+            }
+            assertBalancesMessageReceived()
+        }
+
+        maker2WsClient.apply {
+            assertTradeCreatedMessageReceived(
+                order = limitSellOrder2ApiResponse,
+                price = BigDecimal("17.50"),
+                amount = baseSymbolMakerTradeAmount,
+                fee = AssetAmount(quoteSymbol, "0.00525"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(1, msg.order.executions.size)
+                assertAmount(AssetAmount(baseSymbol, "0.03"), msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "17.500"), msg.order.executions[0].price)
+                assertEquals(ExecutionRole.Maker, msg.order.executions[0].role)
+            }
+            assertBalancesMessageReceived()
+        }
+
+        val trades = getTradesForOrders(listOf(marketSellOrderApiResponse.orderId))
+        assertEquals(2, trades.size)
+
+        takerWsClient.apply {
+            assertLimitsMessageReceived(market.id)
+        }
+
+        makerWsClient.apply {
+            assertLimitsMessageReceived(market.id)
+        }
+
+        assertAmount(AssetAmount(baseSymbol, "0.03"), trades[0].amount)
+        assertAmount(AssetAmount(quoteSymbol, "17.500"), trades[0].price)
+        waitForSettlementToFinish(listOf(trades[0].id.value), SettlementStatus.Failed)
+
+        assertAmount(AssetAmount(baseSymbol, "0.03"), trades[1].amount)
+        assertAmount(AssetAmount(quoteSymbol, "17.500"), trades[1].price)
+        waitForSettlementToFinish(listOf(trades[1].id.value))
+
+        var failedTradeUpdates = 0
+        var completedTradeUpdates = 0
+        takerWsClient.apply {
+            assertMessagesReceived(5) { messages ->
+                println(messages)
+                messages.filter { it.topic == SubscriptionTopic.Trades && it.data is TradeUpdated }.map { it.data as TradeUpdated }.forEach { msg ->
+                    when (msg.trade.settlementStatus) {
+                        SettlementStatus.Failed -> {
+                            failedTradeUpdates++
+                            assertEquals("Insufficient Balance", msg.trade.error)
+                        }
+                        SettlementStatus.Completed -> completedTradeUpdates++
+                        else -> {}
+                    }
+                }
+                assertContainsLimitsMessage(messages, market, base = baseSymbolMakerTradeAmount, quote = AssetAmount(quoteSymbol, "9.4645"))
+            }
+        }
+        assertEquals(1, failedTradeUpdates)
+        assertEquals(1, completedTradeUpdates)
+
+        makerWsClient.apply {
+            assertMessagesReceived(3) { messages ->
+                assertContainsTradeUpdatedMessage(messages) { msg ->
+                    assertEquals(limitSellOrderApiResponse.orderId, msg.trade.orderId)
+                    assertEquals(SettlementStatus.Failed, msg.trade.settlementStatus)
+                    assertEquals("Insufficient Balance", msg.trade.error)
+                }
+                assertContainsBalancesMessage(
+                    messages,
+                    listOf(
+                        ExpectedBalance(baseSymbol, total = AssetAmount(baseSymbol, "0.01"), available = makerStartingBaseBalance - baseWithdrawalAmount),
+                        ExpectedBalance(makerStartingQuoteBalance),
+                    ),
+                )
+                assertContainsLimitsMessage(messages, market, base = makerStartingBaseBalance - baseWithdrawalAmount, quote = makerStartingQuoteBalance)
+            }
+        }
+
+        maker2WsClient.apply {
+            assertMessagesReceived(3) { messages ->
+                assertContainsTradeUpdatedMessage(messages) { msg ->
+                    assertEquals(limitSellOrder2ApiResponse.orderId, msg.trade.orderId)
+                    assertEquals(SettlementStatus.Completed, msg.trade.settlementStatus)
+                }
+                assertContainsBalancesMessage(
+                    messages,
+                    listOf(
+                        ExpectedBalance(baseSymbol, total = maker2StartingBaseBalance - baseSymbolMakerTradeAmount, available = maker2StartingBaseBalance - baseSymbolMakerTradeAmount),
+                        ExpectedBalance(quoteSymbol, total = makerNotionalMinusFee, available = makerNotionalMinusFee),
+                    ),
+                )
+                assertContainsLimitsMessage(messages, market, base = maker2StartingBaseBalance - baseSymbolMakerTradeAmount, quote = makerNotionalMinusFee)
+            }
+        }
+
+        makerWsClient.close()
+        maker2WsClient.close()
+    }
+
+    @Test
+    fun `cross chain settlement failure - all trades in batch fail`() {
+        val market = btcbtc2Market
+        val baseSymbol = btc
+        val quoteSymbol = btc2
+
+        val (takerApiClient, takerWallet, takerWsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.5"),
+                AssetAmount(quoteSymbol, "0.6"),
+            ),
+            deposits = listOf(
+                AssetAmount(baseSymbol, "0.3"),
+                AssetAmount(quoteSymbol, "0.4"),
+            ),
+            subscribeToOrderBook = false,
+            subscribeToOrderPrices = false,
+        )
+
+        val (makerApiClient, makerWallet, makerWsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.5"),
+                AssetAmount(quoteSymbol, "0.7"),
+            ),
+            deposits = listOf(
+                AssetAmount(quoteSymbol, "0.55"),
+            ),
+            subscribeToOrderBook = false,
+            subscribeToOrderPrices = false,
+        )
+
+        // starting onchain balances
+        val makerStartingBaseBalance = makerWallet.getExchangeBalance(baseSymbol)
+        val makerStartingQuoteBalance = makerWallet.getExchangeBalance(quoteSymbol)
+        val takerStartingBaseBalance = takerWallet.getExchangeBalance(baseSymbol)
+        val takerStartingQuoteBalance = takerWallet.getExchangeBalance(quoteSymbol)
+        takerWallet.switchChain(chainIdBySymbol.getValue(baseSymbol.name))
+
+        val baseWithdrawalAmount = AssetAmount(baseSymbol, "0.12")
+
+        val pendingBaseWithdrawal = takerApiClient.createWithdrawal(takerWallet.signWithdraw(baseSymbol.name, baseWithdrawalAmount.inFundamentalUnits)).withdrawal
+        assertEquals(WithdrawalStatus.Pending, pendingBaseWithdrawal.status)
+
+        takerWsClient.apply {
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = takerStartingBaseBalance, available = takerStartingBaseBalance - baseWithdrawalAmount),
+                    ExpectedBalance(quoteSymbol, total = takerStartingQuoteBalance, available = takerStartingQuoteBalance),
+                ),
+            )
+            assertLimitsMessageReceived(market, base = takerStartingBaseBalance - baseWithdrawalAmount, quote = takerStartingQuoteBalance)
+        }
+
+        waitForFinalizedWithdrawal(pendingBaseWithdrawal.id)
+
+        takerWsClient.apply {
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = takerStartingBaseBalance - baseWithdrawalAmount, available = takerStartingBaseBalance - baseWithdrawalAmount),
+                    ExpectedBalance(quoteSymbol, total = takerStartingQuoteBalance, available = takerStartingQuoteBalance),
+                ),
+            )
+        }
+
+        // hack to resubmit the same transaction again bypassing the sequencer. This way taker will not have a
+        // sufficient on chain balance for the order so settlement will fail.
+        transaction {
+            WithdrawalEntity[pendingBaseWithdrawal.id].status = WithdrawalStatus.Sequenced
+        }
+
+        waitForFinalizedWithdrawal(pendingBaseWithdrawal.id)
+
+        takerWsClient.apply {
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = takerStartingBaseBalance - baseWithdrawalAmount * BigDecimal(2), available = takerStartingBaseBalance - baseWithdrawalAmount),
+                    ExpectedBalance(quoteSymbol, total = takerStartingQuoteBalance, available = takerStartingQuoteBalance),
+                ),
+            )
+        }
+
+        // place a limit order
+        val limitBuyOrderApiResponse = makerApiClient.createLimitOrder(
+            market,
+            OrderSide.Buy,
+            amount = BigDecimal("0.08"),
+            price = BigDecimal("0.999"),
+            makerWallet,
+        )
+        makerWsClient.apply {
+            assertLimitOrderCreatedMessageReceived(limitBuyOrderApiResponse)
+            assertLimitsMessageReceived(market, base = BigDecimal("0"), quote = BigDecimal("0.47008"))
+        }
+
+        // place a sell order
+        val marketBuyOrderApiResponse = takerApiClient.createMarketOrder(
+            market,
+            OrderSide.Sell,
+            BigDecimal("0.08"),
+            takerWallet,
+        )
+
+        takerWsClient.apply {
+            assertMarketOrderCreatedMessageReceived(marketBuyOrderApiResponse)
+            assertTradeCreatedMessageReceived(
+                order = marketBuyOrderApiResponse,
+                price = BigDecimal("0.999"),
+                amount = AssetAmount(baseSymbol, marketBuyOrderApiResponse.order.amount),
+                fee = AssetAmount(quoteSymbol, "0.0015984"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(1, msg.order.executions.size)
+                assertAmount(AssetAmount(baseSymbol, "0.08"), msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "0.999"), msg.order.executions[0].price)
+                assertEquals(ExecutionRole.Taker, msg.order.executions[0].role)
+            }
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = BigDecimal("0.06"), available = BigDecimal("0.1")),
+                    ExpectedBalance(quoteSymbol, total = BigDecimal("0.4"), available = BigDecimal("0.4783216")),
+                ),
+            )
+        }
+
+        makerWsClient.apply {
+            assertTradeCreatedMessageReceived(
+                order = limitBuyOrderApiResponse,
+                price = BigDecimal("0.999"),
+                amount = AssetAmount(baseSymbol, "0.08"),
+                fee = AssetAmount(quoteSymbol, "0.0007992"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(1, msg.order.executions.size)
+                assertAmount(AssetAmount(baseSymbol, "0.08"), msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "0.999"), msg.order.executions[0].price)
+                assertEquals(ExecutionRole.Maker, msg.order.executions[0].role)
+            }
+            assertBalancesMessageReceived(
+                listOf(
+                    ExpectedBalance(baseSymbol, total = BigDecimal("0"), available = BigDecimal("0.08")),
+                    ExpectedBalance(quoteSymbol, total = BigDecimal("0.55"), available = BigDecimal("0.4692808")),
+                ),
+            )
+        }
+
+        val trade = getTradesForOrders(listOf(marketBuyOrderApiResponse.orderId)).first()
+
+        takerWsClient.apply {
+            assertLimitsMessageReceived(market, base = BigDecimal("0.1"), quote = BigDecimal("0.4783216"))
+        }
+
+        makerWsClient.apply {
+            assertLimitsMessageReceived(market, base = BigDecimal("0.08"), quote = BigDecimal("0.4692808"))
+        }
+
+        assertAmount(AssetAmount(baseSymbol, "0.08"), trade.amount)
+        assertAmount(AssetAmount(quoteSymbol, "0.999"), trade.price)
+
+        waitForSettlementToFinish(listOf(trade.id.value), SettlementStatus.Failed)
+
+        takerWsClient.apply {
+            assertMessagesReceived(3) { messages ->
+                assertContainsTradeUpdatedMessage(messages) { msg ->
+                    assertEquals(marketBuyOrderApiResponse.orderId, msg.trade.orderId)
+                    assertEquals(SettlementStatus.Failed, msg.trade.settlementStatus)
+                    assertEquals("Insufficient Balance", msg.trade.error)
+                }
+                assertContainsBalancesMessage(
+                    messages,
+                    listOf(
+                        ExpectedBalance(baseSymbol, total = AssetAmount(baseSymbol, "0.06"), available = takerStartingBaseBalance - baseWithdrawalAmount),
+                        ExpectedBalance(takerStartingQuoteBalance),
+                    ),
+                )
+                assertContainsLimitsMessage(messages, market, base = takerStartingBaseBalance - baseWithdrawalAmount, quote = takerStartingQuoteBalance)
+            }
+        }
+
+        makerWsClient.apply {
+            assertMessagesReceived(3) { messages ->
+                assertContainsTradeUpdatedMessage(messages) { msg ->
+                    assertEquals(limitBuyOrderApiResponse.orderId, msg.trade.orderId)
+                    assertEquals(SettlementStatus.Failed, msg.trade.settlementStatus)
+                    assertEquals("Insufficient Balance", msg.trade.error)
+                }
+                assertContainsBalancesMessage(
+                    messages,
+                    listOf(
+                        ExpectedBalance(makerStartingBaseBalance),
+                        ExpectedBalance(makerStartingQuoteBalance),
+                    ),
+                )
+                assertContainsLimitsMessage(messages, market, base = makerStartingBaseBalance, quote = makerStartingQuoteBalance)
+            }
+        }
+
+        makerWsClient.close()
+
+        waitForSettlementBatchToFinish()
+    }
+
+    @Test
+    fun `cross chain settlement failure - some trades in batch fail, some succeed`() {
+        val market = btcbtc2Market
+        val baseSymbol = btc
+        val quoteSymbol = btc2
+
+        val (takerApiClient, takerWallet, takerWsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.22"),
+                AssetAmount(quoteSymbol, "0.2"),
+            ),
+            deposits = listOf(
+                AssetAmount(quoteSymbol, "0.1"),
+            ),
+            subscribeToOrderBook = false,
+            subscribeToOrderPrices = false,
+        )
+
+        val (makerApiClient, makerWallet, makerWsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.1"),
+            ),
+            deposits = listOf(
+                AssetAmount(baseSymbol, "0.05"),
+            ),
+            subscribeToOrderBook = false,
+            subscribeToOrderPrices = false,
+        )
+
+        val (maker2ApiClient, maker2Wallet, maker2WsClient) = setupTrader(
+            market.id,
+            airdrops = listOf(
+                AssetAmount(baseSymbol, "0.1"),
+            ),
+            deposits = listOf(
+                AssetAmount(baseSymbol, "0.03"),
+            ),
+            subscribeToOrderBook = false,
+            subscribeToOrderPrices = false,
+        )
+
+        // starting onchain balances
+        val makerStartingBaseBalance = makerWallet.getExchangeBalance(baseSymbol)
+        val makerStartingQuoteBalance = makerWallet.getExchangeBalance(quoteSymbol)
+        val maker2StartingBaseBalance = maker2Wallet.getExchangeBalance(baseSymbol)
+        val baseSymbolMakerTradeAmount = AssetAmount(baseSymbol, "0.03")
+        val makerNotionalMinusFee = AssetAmount(quoteSymbol, "0.0296703")
+
+        val baseWithdrawalAmount = AssetAmount(baseSymbol, "0.02")
+
+        val pendingBaseWithdrawal = makerApiClient.createWithdrawal(makerWallet.signWithdraw(baseSymbol.name, baseWithdrawalAmount.inFundamentalUnits)).withdrawal
+        assertEquals(WithdrawalStatus.Pending, pendingBaseWithdrawal.status)
+
+        makerWsClient.apply {
+            assertBalancesMessageReceived()
+            assertLimitsMessageReceived(market, base = makerStartingBaseBalance - baseWithdrawalAmount, quote = makerStartingQuoteBalance)
+        }
+
+        waitForFinalizedWithdrawal(pendingBaseWithdrawal.id)
+
+        makerWsClient.apply {
+            assertBalancesMessageReceived()
+        }
+
+        // hack to resubmit the same transaction again bypassing the sequencer. This way taker will not have a
+        // sufficient on chain balance for the order so settlement will fail.
+        transaction {
+            WithdrawalEntity[pendingBaseWithdrawal.id].status = WithdrawalStatus.Sequenced
+        }
+
+        waitForFinalizedWithdrawal(pendingBaseWithdrawal.id)
+
+        makerWsClient.apply {
+            assertBalancesMessageReceived()
+        }
+
+        // place a limit order
+        val limitSellOrderApiResponse = makerApiClient.createLimitOrder(
+            market,
+            OrderSide.Sell,
+            amount = BigDecimal("0.03"),
+            price = BigDecimal("0.999"),
+            makerWallet,
+        )
+        makerWsClient.apply {
+            assertLimitOrderCreatedMessageReceived(limitSellOrderApiResponse)
+            assertLimitsMessageReceived(market.id)
+        }
+
+        val limitSellOrder2ApiResponse = maker2ApiClient.createLimitOrder(
+            market,
+            OrderSide.Sell,
+            amount = BigDecimal("0.03"),
+            price = BigDecimal("0.999"),
+            maker2Wallet,
+        )
+        maker2WsClient.apply {
+            assertLimitOrderCreatedMessageReceived(limitSellOrder2ApiResponse)
+            assertLimitsMessageReceived(market.id)
+        }
+
+        // place a sell order
+        val marketSellOrderApiResponse = takerApiClient.createMarketOrder(
+            market,
+            OrderSide.Buy,
+            BigDecimal("0.06"),
+            takerWallet,
+        )
+
+        takerWsClient.apply {
+            assertMarketOrderCreatedMessageReceived(marketSellOrderApiResponse)
+            assertTradeCreatedMessageReceived(
+                order = marketSellOrderApiResponse,
+                price = BigDecimal("0.999"),
+                amount = baseSymbolMakerTradeAmount,
+                fee = AssetAmount(quoteSymbol, "0.0005994"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertTradeCreatedMessageReceived(
+                order = marketSellOrderApiResponse,
+                price = BigDecimal("0.999"),
+                amount = baseSymbolMakerTradeAmount,
+                fee = AssetAmount(quoteSymbol, "0.0005994"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(2, msg.order.executions.size)
+                assertAmount(baseSymbolMakerTradeAmount, msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "0.999"), msg.order.executions[0].price)
+                assertAmount(baseSymbolMakerTradeAmount, msg.order.executions[1].amount)
+                assertAmount(AssetAmount(quoteSymbol, "0.999"), msg.order.executions[1].price)
+                assertEquals(ExecutionRole.Taker, msg.order.executions[0].role)
+                assertEquals(ExecutionRole.Taker, msg.order.executions[1].role)
+            }
+            assertBalancesMessageReceived()
+        }
+
+        makerWsClient.apply {
+            assertTradeCreatedMessageReceived(
+                order = limitSellOrderApiResponse,
+                price = BigDecimal("0.999"),
+                amount = baseSymbolMakerTradeAmount,
+                fee = AssetAmount(quoteSymbol, "0.0002997"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(1, msg.order.executions.size)
+                assertAmount(baseSymbolMakerTradeAmount, msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "0.999"), msg.order.executions[0].price)
+                assertEquals(ExecutionRole.Maker, msg.order.executions[0].role)
+            }
+            assertBalancesMessageReceived()
+        }
+
+        maker2WsClient.apply {
+            assertTradeCreatedMessageReceived(
+                order = limitSellOrder2ApiResponse,
+                price = BigDecimal("0.999"),
+                amount = baseSymbolMakerTradeAmount,
+                fee = AssetAmount(quoteSymbol, "0.0002997"),
+                settlementStatus = SettlementStatus.Pending,
+            )
+            assertOrderUpdatedMessageReceived { msg ->
+                assertEquals(OrderStatus.Filled, msg.order.status)
+                assertEquals(1, msg.order.executions.size)
+                assertAmount(AssetAmount(baseSymbol, "0.03"), msg.order.executions[0].amount)
+                assertAmount(AssetAmount(quoteSymbol, "0.999"), msg.order.executions[0].price)
+                assertEquals(ExecutionRole.Maker, msg.order.executions[0].role)
+            }
+            assertBalancesMessageReceived()
+        }
+
+        val trades = getTradesForOrders(listOf(marketSellOrderApiResponse.orderId))
+        assertEquals(2, trades.size)
+
+        takerWsClient.apply {
+            assertLimitsMessageReceived(market.id)
+        }
+
+        makerWsClient.apply {
+            assertLimitsMessageReceived(market.id)
+        }
+
+        assertAmount(AssetAmount(baseSymbol, "0.03"), trades[0].amount)
+        assertAmount(AssetAmount(quoteSymbol, "0.999"), trades[0].price)
+        waitForSettlementToFinish(listOf(trades[0].id.value), SettlementStatus.Failed)
+
+        assertAmount(AssetAmount(baseSymbol, "0.03"), trades[1].amount)
+        assertAmount(AssetAmount(quoteSymbol, "0.999"), trades[1].price)
+        waitForSettlementToFinish(listOf(trades[1].id.value))
+
+        var failedTradeUpdates = 0
+        var completedTradeUpdates = 0
+        takerWsClient.apply {
+            assertMessagesReceived(5) { messages ->
+                println(messages)
+                messages.filter { it.topic == SubscriptionTopic.Trades && it.data is TradeUpdated }.map { it.data as TradeUpdated }.forEach { msg ->
+                    when (msg.trade.settlementStatus) {
+                        SettlementStatus.Failed -> {
+                            failedTradeUpdates++
+                            assertEquals("Insufficient Balance", msg.trade.error)
+                        }
+                        SettlementStatus.Completed -> completedTradeUpdates++
+                        else -> {}
+                    }
+                }
+                assertContainsLimitsMessage(messages, market, base = baseSymbolMakerTradeAmount, quote = AssetAmount(quoteSymbol, "0.0694306"))
+            }
+        }
+        assertEquals(1, failedTradeUpdates)
+        assertEquals(1, completedTradeUpdates)
+
+        makerWsClient.apply {
+            assertMessagesReceived(3) { messages ->
+                assertContainsTradeUpdatedMessage(messages) { msg ->
+                    assertEquals(limitSellOrderApiResponse.orderId, msg.trade.orderId)
+                    assertEquals(SettlementStatus.Failed, msg.trade.settlementStatus)
+                    assertEquals("Insufficient Balance", msg.trade.error)
+                }
+                assertContainsBalancesMessage(
+                    messages,
+                    listOf(
+                        ExpectedBalance(baseSymbol, total = AssetAmount(baseSymbol, "0.01"), available = makerStartingBaseBalance - baseWithdrawalAmount),
+                        ExpectedBalance(makerStartingQuoteBalance),
+                    ),
+                )
+                assertContainsLimitsMessage(messages, market, base = makerStartingBaseBalance - baseWithdrawalAmount, quote = makerStartingQuoteBalance)
+            }
+        }
+
+        maker2WsClient.apply {
+            assertMessagesReceived(3) { messages ->
+                assertContainsTradeUpdatedMessage(messages) { msg ->
+                    assertEquals(limitSellOrder2ApiResponse.orderId, msg.trade.orderId)
+                    assertEquals(SettlementStatus.Completed, msg.trade.settlementStatus)
+                }
+                assertContainsBalancesMessage(
+                    messages,
+                    listOf(
+                        ExpectedBalance(baseSymbol, total = maker2StartingBaseBalance - baseSymbolMakerTradeAmount, available = maker2StartingBaseBalance - baseSymbolMakerTradeAmount),
+                        ExpectedBalance(quoteSymbol, total = makerNotionalMinusFee, available = makerNotionalMinusFee),
+                    ),
+                )
+                assertContainsLimitsMessage(messages, market, base = maker2StartingBaseBalance - baseSymbolMakerTradeAmount, quote = makerNotionalMinusFee)
+            }
+        }
+
+        makerWsClient.close()
+        maker2WsClient.close()
+        waitForSettlementBatchToFinish()
+    }
+}
